backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
supabase==2.0.0
python-multipart==0.0.6
aiofiles==23.2.1
pydantic>=2.9.0
requests==2.31.0
resend==2.1.0

